* Refactored data mid layers * Added exception handling decoupling between REST Controller and Database * Some database/hibernate performance improvements * Added initial container for todo app * Added pg admin for debugging purposes

+++ src/main/java/com/github/deberth/todo/core/TodoServiceDatabase.java
@@ -3,12 +3,10 @@ package com.github.deberth.todo.core;
 import com.github.deberth.todo.api.Todo;
 import com.github.deberth.todo.db.TaskDAO;
 import com.github.deberth.todo.db.TodoDAO;
-import io.dropwizard.hibernate.UnitOfWork;
 
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
 import java.util.List;
-import java.util.stream.Collectors;
 
 public class TodoServiceDatabase extends TodoService{
 
@@ -17,33 +15,65 @@ public class TodoServiceDatabase extends TodoService{
     }
 
     @Override
-    public List<Todo> findAllTodos() {
-        return this.todoDAO.findAll();
+    public TodoServiceResponse findAllTodos() {
+        return new TodoServiceResponse(OK, this.todoDAO.findAll());
     }
 
     @Override
-    public Todo findTodoById(Integer id) {
-        return this.todoDAO.find(id);
+    public TodoServiceResponse findTodoById(Integer id) {
+        Todo found = this.todoDAO.find(id);
+        if (found == null) {
+            return new TodoServiceResponse(NOT_FOUND, null);
+        }
+        return new TodoServiceResponse(OK, found);
     }
 
     @Override
-    public Todo createNewTodo(@NotNull @Valid Todo todo) {
-        todo.setTasks(todo.getTasks().stream().map(t -> this.taskDAO.create(t)).collect(Collectors.toList()));
-        return this.todoDAO.create(todo);
+    public TodoServiceResponse createNewTodo(@NotNull @Valid Todo todo) {
+
+        // Avoid invalid POST with already existing ID
+        if (todo.getId() != null) {
+            Todo found = this.todoDAO.find(todo.getId());
+            if (found == null) {
+                return new TodoServiceResponse(CONFLICT);
+            }
+        }
+
+        Todo createdTodo = this.todoDAO.create(todo);
+        return new TodoServiceResponse(CREATED, createdTodo);
     }
 
     @Override
-    public void removeTodo(Integer id) {
+    public TodoServiceResponse removeTodo(Integer id) {
         Todo todo = this.todoDAO.find(id);
-        todo.getTasks().forEach(t -> this.taskDAO.remove(t.getId()));
-        this.todoDAO.remove(id);
+        if (todo != null) {
+            this.todoDAO.remove(id);
+            return new TodoServiceResponse(NO_CONTENT, null);
+        } else {
+            return new TodoServiceResponse(NOT_FOUND, null);
+        }
     }
 
     @Override
-    public void updateTodo(@NotNull @Valid Todo todo) {
-        Todo found = this.todoDAO.find(todo.getId());
-        todo.getTasks().forEach(t -> this.taskDAO.remove(t.getId()));
-        this.todoDAO.update(todo.getId(), todo);
+    public TodoServiceResponse updateTodo(Integer id, @NotNull @Valid Todo updatedTodo) {
+        // Check for invalid differences between path id and json body id
+        if (updatedTodo.getId() != null && updatedTodo.getId() != id) {
+            return new TodoServiceResponse(BAD_REQUEST, null);
+        }
+        Todo foundTodo = this.todoDAO.find(id);
+        if(foundTodo == null) {
+            return createNewTodo(updatedTodo);
+        }
+
+        // Update found todo with updated information
+        foundTodo.setName(updatedTodo.getName());
+        foundTodo.setDescription(updatedTodo.getDescription());
+        // Avoid detached tasks exception
+        foundTodo.getTasks().clear();
+        foundTodo.getTasks().addAll(updatedTodo.getTasks());
+
+        this.todoDAO.update(foundTodo.getId(), foundTodo);
+        return new TodoServiceResponse(NO_CONTENT, null);
     }
